Move hooks -> clp
