+++ ui/core/src/index.ts
@@ -2,6 +2,6 @@ export * from "./entities";
 export * from "./services";
 export * from "./usecases";
 export * from "./store";
-export * from "./hooks";
+export * from "./usecases/clp/calculators";
 export * from "./config";
 export * from "./utils";
