+++ ui/core/src/usecases/clp/calculators/useField.ts
@@ -1,6 +1,6 @@
 import { computed, Ref } from "@vue/reactivity";
-import { Asset, AssetAmount } from "../entities";
-import { toBaseUnits } from "../utils";
+import { Asset, AssetAmount } from "../../../entities";
+import { toBaseUnits } from "../../../utils";
 
 export function useField(amount: Ref<string>, symbol: Ref<string | null>) {
   const asset = computed(() => {
