+++ ui/core/src/usecases/clp/calculators/utils.test.ts
@@ -1,5 +1,5 @@
-import { AssetAmount, Pool, Network, Asset } from "../entities";
-import { getTestingTokens } from "../test/utils/getTestingToken";
+import { AssetAmount, Pool, Network, Asset } from "../../../entities";
+import { getTestingTokens } from "../../../test/utils/getTestingToken";
 import { assetPriceMessage } from "./utils";
 
 describe("assets with decimals", () => {
